refactor: 패키지명을 고친다
 - 기존 패키지 구조는 domain 패키지가 중복된다. 한쪽에 domain 대신
   entity를 쓰도록 고친다
