+++ src/main/java/org/swyp/weddy/domain/wiki/dao/WikiMapper.java
@@ -1,7 +1,7 @@
 package org.swyp.weddy.domain.wiki.dao;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.swyp.weddy.domain.wiki.domain.Wiki;
+import org.swyp.weddy.domain.wiki.entity.Wiki;
 
 @Mapper
 public interface WikiMapper {
