+++ src/main/java/org/swyp/weddy/domain/wiki/service/WikiService.java
@@ -3,7 +3,7 @@ package org.swyp.weddy.domain.wiki.service;
 import org.springframework.stereotype.Service;
 import org.swyp.weddy.common.exception.ErrorCode;
 import org.swyp.weddy.domain.wiki.dao.WikiMapper;
-import org.swyp.weddy.domain.wiki.domain.Wiki;
+import org.swyp.weddy.domain.wiki.entity.Wiki;
 import org.swyp.weddy.domain.wiki.exception.WikiNotFoundException;
 import org.swyp.weddy.domain.wiki.service.dto.WikiDto;
 import org.swyp.weddy.domain.wiki.web.response.WikiResponse;
